+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## [Unreleased]
 
+
+## [0.1.4]
+* Add `close_current` method for depth first traversals.
+
+### Added
+* Add `close_current` method for depth first traversals.
+    + Add methods:
+        - `traverse::DftEvent::<Node<T>>::close_current`
+        - `traverse::DftEvent::<FrozenNode<T>>::close_current`
+        - `traverse::DftEvent::<HotNode<T>>::close_current`
+        - `traverse::DepthFirstTraverser::close_current`
+        - `traverse::StableDepthFirstTraverser::close_current`
+        - `traverse::StableShallowDepthFirstTraverser::close_current`
+
+
 ## [0.1.3]
 * Implement `Clone` for `Tree<T>` and `TreeWeak<T>`.
 
@@ -89,7 +104,8 @@
 
 Initial release.
 
-[Unreleased]: <https://gitlab.com/nop_thread/dendron/-/compare/v0.1.3...develop>
+[Unreleased]: <https://gitlab.com/nop_thread/dendron/-/compare/v0.1.4...develop>
+[0.1.4]: <https://gitlab.com/nop_thread/dendron/-/tags/v0.1.4>
 [0.1.3]: <https://gitlab.com/nop_thread/dendron/-/tags/v0.1.3>
 [0.1.2]: <https://gitlab.com/nop_thread/dendron/-/tags/v0.1.2>
 [0.1.1]: <https://gitlab.com/nop_thread/dendron/-/tags/v0.1.1>